feat(web): one-click dev sign-in button for local development (#4778)
* feat(web): one-click dev sign-in button for local development

Replaces `DevAuthForm` (form with prefilled email + password) on the
web sign-in page with a single "Sign in as Local Admin (dev)" button.
Dev creds never change in local profile, so the inputs were noise.

The button calls authClient.signIn.email and, on first run, falls
back to authClient.signUp.email then retries — mirroring
apps/desktop LocalDevAuthForm behavior. Gated by
NODE_ENV !== "production" so it tree-shakes out of prod builds.
Removed from the sign-up page (redundant with the sign-in
auto-create).

Also fills three required client-side env vars in .env.example so
apps/web's t3-env validation passes on a fresh clone. Without these,
validation throws during module init and silently kills React
hydration — every sign-in button (OAuth + dev) becomes a
no-network no-op.

  NEXT_PUBLIC_POSTHOG_KEY=phc_local_dev_disabled
  NEXT_PUBLIC_POSTHOG_HOST=https://us.posthog.com
  NEXT_PUBLIC_RELAY_URL=http://localhost:4734

shouldSkipEnvValidation() skips server-side validation in the local
profile, but Next.js statically inlines NEXT_PUBLIC_* into the
client bundle and t3-env still revalidates there.

.gitignore also picks up `.claude/worktrees/` and `.spec/` so
per-developer artifacts stop leaking into PRs.

* refactor: extract shared DevSignInButton UI + dev credentials

Replaces apps/desktop's LocalDevAuthForm (email/password form) with a
one-click button matching the web. Both apps now wrap a single shared
UI component and supply only platform-specific glue:

  packages/ui/.../dev-sign-in-button.tsx   pure UI button + state
  packages/shared/dev-credentials.ts       DEV_EMAIL/PASSWORD/NAME

  apps/web/.../DevSignInButton.tsx         Next.js useRouter +
                                           authClient.signIn.email
  apps/desktop/.../LocalDevSignInButton/   TanStack useNavigate +
                                           raw fetch + electron-trpc
                                           token persistence

Eliminates dev-cred duplication across four files (db seed, desktop
main auto-sign-in, desktop renderer wrapper, web wrapper). The
desktop sign-in still uses raw fetch + saveToken because Electron
manages tokens itself; the web wrapper uses cookies via
@superset/auth/client. Routing diverges naturally between the two
routers, which is why the shared piece is just the button itself.

Dev-only on both platforms — desktop gates on SUPERSET_PROFILE=local,
web gates on NODE_ENV !== "production".

* fix(web): restore DevSignInButton on sign-up page

The button auto-creates the seeded account on first run via its
sign-up fallback, so the sign-up route gives the same one-click
local-dev path as sign-in. Gated by NODE_ENV !== "production".
