+++ .env.example
@@ -91,7 +91,8 @@ BLOB_READ_WRITE_TOKEN=
 # -----------------------------------------------------------------------------
 # PostHog Analytics (Local Dev)
 # -----------------------------------------------------------------------------
-NEXT_PUBLIC_POSTHOG_KEY=
+NEXT_PUBLIC_POSTHOG_KEY=phc_local_dev_disabled
+NEXT_PUBLIC_POSTHOG_HOST=https://us.posthog.com
 POSTHOG_API_KEY=
 POSTHOG_PROJECT_ID=
 
@@ -146,5 +147,5 @@ SUPERSET_MCP_API_KEY=
 # Relay service URL (the v2 tunnel proxy that forwards cloud API calls
 # to host-service instances on user devices). Local dev: http://localhost:4734
 RELAY_URL=
-# Browser-exposed relay URL — the web app's host-service tRPC + terminal WS.
-NEXT_PUBLIC_RELAY_URL=
+# Browser-exposed relay URL for the web app's host-service tRPC + terminal WS.
+NEXT_PUBLIC_RELAY_URL=http://localhost:4734
