--- apps/desktop/src/renderer/routes/sign-in/components/LocalDevAuthForm/index.ts
@@ -1,1 +1,0 @@
-export { LocalDevAuthForm } from "./LocalDevAuthForm";
