--- apps/web/src/app/(auth)/components/DevAuthForm/index.ts
@@ -1,1 +1,0 @@
-export { DevAuthForm } from "./DevAuthForm";
