+++ apps/web/src/app/(auth)/components/DevSignInButton/index.ts
@@ -1,0 +1,1 @@
+export { DevSignInButton } from "./DevSignInButton";
