+++ apps/web/src/app/(auth)/sign-in/[[...sign-in]]/page.tsx
@@ -8,7 +8,7 @@ import { useState } from "react";
 import { FaGithub } from "react-icons/fa";
 import { FcGoogle } from "react-icons/fc";
 import { env } from "@/env";
-import { DevAuthForm } from "../../components/DevAuthForm";
+import { DevSignInButton } from "../../components/DevSignInButton";
 
 const isDev = process.env.NODE_ENV !== "production";
 
@@ -69,7 +69,7 @@ export default function SignInPage() {
 				{error && (
 					<p className="text-destructive text-center text-sm">{error}</p>
 				)}
-				{isDev && <DevAuthForm mode="sign-in" callbackURL={callbackURL} />}
+				{isDev && <DevSignInButton callbackURL={callbackURL} />}
 				<Button
 					variant="outline"
 					disabled={isLoading}
