+++ apps/web/src/app/(auth)/sign-up/[[...sign-up]]/page.tsx
@@ -7,7 +7,7 @@ import { useState } from "react";
 import { FaGithub } from "react-icons/fa";
 import { FcGoogle } from "react-icons/fc";
 import { env } from "@/env";
-import { DevAuthForm } from "../../components/DevAuthForm";
+import { DevSignInButton } from "../../components/DevSignInButton";
 
 const isDev = process.env.NODE_ENV !== "production";
 
@@ -64,9 +64,7 @@ export default function SignUpPage() {
 				{error && (
 					<p className="text-destructive text-center text-sm">{error}</p>
 				)}
-				{isDev && (
-					<DevAuthForm mode="sign-up" callbackURL={env.NEXT_PUBLIC_WEB_URL} />
-				)}
+				{isDev && <DevSignInButton callbackURL={env.NEXT_PUBLIC_WEB_URL} />}
 				<Button
 					variant="outline"
 					disabled={isLoading}
