+++ packages/shared/src/dev-credentials.ts
@@ -1,0 +1,11 @@
+/**
+ * Seeded local-dev account used by:
+ * - packages/db seed-dev to insert the user
+ * - apps/desktop LocalDevAuthForm + dev-auto-sign-in to authenticate against it
+ * - apps/web DevSignInButton to authenticate against it
+ *
+ * Only valid in the `local` deployment profile. Never use in production.
+ */
+export const DEV_EMAIL = "admin@local.test";
+export const DEV_PASSWORD = "supersetdev";
+export const DEV_NAME = "Local Admin";
